requirements.txt
Flask==2.2.2
Flask_SocketIO==5.3.1
keras==2.10.0
numpy==1.23.3
opencv_python_headless==4.6.0.66
Pillow==9.3.0
tensorflow==2.10.0
tensorflow_gpu==2.10.0
